separate file for ISupportInitialize.
